Keep pushing stuff inside the fixture
