+++ catalogue_pipeline/merger/src/main/scala/uk/ac/wellcome/platform/merger/Main.scala
@@ -1,10 +1,12 @@
 package uk.ac.wellcome.platform.merger
 
+import akka.actor.ActorSystem
 import com.typesafe.config.{Config, ConfigFactory}
 import grizzled.slf4j.Logging
 import uk.ac.wellcome.config.core.builders.AkkaBuilder
 import uk.ac.wellcome.config.messaging.builders.{MessagingBuilder, SQSBuilder}
 import uk.ac.wellcome.config.storage.builders.VHSBuilder
+import uk.ac.wellcome.json.JsonUtil._
 import uk.ac.wellcome.messaging.sns.NotificationMessage
 import uk.ac.wellcome.models.work.internal.{BaseWork, TransformedBaseWork}
 import uk.ac.wellcome.platform.merger.services.{Merger, MergerManager, MergerWorkerService, RecorderPlaybackService}
@@ -16,6 +18,8 @@ import scala.concurrent.duration.Duration
 object Main extends App with Logging {
   val config: Config = ConfigFactory.load()
 
+  implicit val actorSystem: ActorSystem =
+    AkkaBuilder.buildActorSystem()
   implicit val executionContext: ExecutionContext =
     AkkaBuilder.buildExecutionContext()
