+++ catalogue_pipeline/merger/src/test/scala/uk/ac/wellcome/platform/merger/fixtures/WorkerServiceFixture.scala
@@ -4,20 +4,13 @@ import uk.ac.wellcome.messaging.sns.NotificationMessage
 import uk.ac.wellcome.messaging.test.fixtures.Messaging
 import uk.ac.wellcome.messaging.test.fixtures.SNS.Topic
 import uk.ac.wellcome.messaging.test.fixtures.SQS.Queue
-import uk.ac.wellcome.models.work.internal.{BaseWork, TransformedBaseWork}
+import uk.ac.wellcome.models.work.internal.BaseWork
 import uk.ac.wellcome.monitoring.MetricsSender
 import uk.ac.wellcome.platform.merger.services.{Merger, MergerManager, MergerWorkerService, RecorderPlaybackService}
-import uk.ac.wellcome.storage.ObjectStore
 import uk.ac.wellcome.storage.fixtures.S3
-import uk.ac.wellcome.storage.vhs.{EmptyMetadata, VersionedHybridStore}
 import uk.ac.wellcome.test.fixtures.TestWith
 
-trait WorkerServiceFixture extends Messaging with S3 {
-  type TransformedBaseWorkVHS = VersionedHybridStore[
-    TransformedBaseWork,
-    EmptyMetadata,
-    ObjectStore[TransformedBaseWork]]
-
+trait WorkerServiceFixture extends LocalWorksVhs with Messaging with S3 {
   def withWorkerService[R](
     vhs: TransformedBaseWorkVHS,
     topic: Topic,
